arg for redispubsub sh scripts for build dir

+++ redispubsub_stop.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 # Run this for stopping sample nmtoken network on multiple nodes
+if [ -z "$1" ]; then
+  echo "Usage: $0 <cmake build dir>"
+  exit 1
+fi
 
-kill $(ps -ef | grep [.]/build/client | awk 'BEGIN{}{if (FNR==NR) {codes[$2]=$2;} else{next}}END{i=0;for (key in codes){i++;if (i<=6 && i<=length(codes)) {print codes[key]}}}')
+kill $(ps -ef | grep [.]/$bdir/client | awk 'BEGIN{}{if (FNR==NR) {codes[$2]=$2;} else{next}}END{i=0;for (key in codes){i++;if (i<=6 && i<=length(codes)) {print codes[key]}}}')
 (docker compose down)
